ifenslave: wants net/if.h

+++ networking/ifenslave.c
@@ -100,14 +100,15 @@
 
 #include "libbb.h"
 
+#include <net/if.h>
 #include <net/if_arp.h>
 #include <linux/if_bonding.h>
 #include <linux/sockios.h>
 
-typedef unsigned long long u64; /* hack, so we may include kernel's ethtool.h */
-typedef uint32_t u32;           /* ditto */
-typedef uint16_t u16;           /* ditto */
-typedef uint8_t u8;             /* ditto */
+typedef uint64_t u64; /* hack, so we may include kernel's ethtool.h */
+typedef uint32_t u32; /* ditto */
+typedef uint16_t u16; /* ditto */
+typedef uint8_t u8;   /* ditto */
 #include <linux/ethtool.h>
